requirements.txt
python-telegram-bot==20.8
pymongo[srv]
flask
dnspython